+++ Reddit/Intermediate/possible_pies/src/lib.rs
@@ -1,4 +1,6 @@
 
-pub fn number_of_pies(pumpkin: i32, apples: i32, eggs: i32, milk: i32, sugar: i32) -> (i32, i32) {
+pub fn number_of_pies(pumpkin: i32, apples: i32, eggs: i32,
+                      milk: i32, sugar: i32) -> (i32, i32) {
+
     (pumpkin+apples, eggs+milk+sugar)
 }