+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 1.05.2018
+
+* [Python] Backwards Read Primes - Test setup (CodeWars)
+
 #### 18.04.2018
 
 * [Rust] Possible pies - Test setup (Reddit)